+++ client_test.go
@@ -126,7 +126,7 @@ func TestParseGetLegislatorsJSON(t *testing.T) {
 
 func TestParseMemberPFDJSON(t *testing.T) {
 	t.Run("Correctly parses valid JSON", func(t *testing.T) {
-		json, err := ioutil.ReadFile("mockPFDResponse.json")
+		json, err := ioutil.ReadFile("mocks/mockPFDResponse.json")
 		if err != nil {
 			t.Fatalf("Error reading mock data from file: %s", err.Error())
 		}
@@ -169,35 +169,6 @@ func TestParseMemberPFDJSON(t *testing.T) {
 		wantedErrorMessage := unable_to_parse_error_message
 		assertErrorMessage(err, wantedErrorMessage, t)
 	})
-}
-
-func assertErrorExists(err error, t *testing.T) {
-	t.Helper()
-	if err == nil {
-		t.Fatalf("Wanted error but got nil")
-	}
-}
-
-func assertErrorMessage(err error, expectedMessage string, t *testing.T) {
-	t.Helper()
-	if err.Error() != expectedMessage {
-		t.Fatalf("Wanted error message %s but got %s", expectedMessage, err.Error())
-	}
-
-}
-
-func assertStringMatches(got, wanted string, t *testing.T) {
-	t.Helper()
-	if got != wanted {
-		t.Fatalf("Got string %s wanted string %s", got, wanted)
-	}
-}
-
-func assertSliceLength(got, wanted int, t *testing.T) {
-	t.Helper()
-	if got != wanted {
-		t.Fatalf("Got slice length %d wanted %d", got, wanted)
-	}
 }
 
 func buildMockResponse(statusCode int, jsonBody string) http.Response {
